+++ vms/platformvm/block/executor/block.go
@@ -89,7 +89,7 @@ func (b *Block) Timestamp() time.Time {
 func (b *Block) Options(context.Context) ([2]snowman.Block, error) {
 	options := options{
 		log:                     b.manager.ctx.Log,
-		primaryUptimePercentage: b.manager.txExecutorBackend.Config.UptimePercentage,
+		primaryUptimePercentage: b.manager.txExecutorBackend.Config.RequiredUptime,
 		uptimes:                 b.manager.txExecutorBackend.Uptimes,
 		state:                   b.manager.backend.state,
 	}